@@ -17,7 +17,7 @@ license-files = [ "LICENSE" ]
 authors = [
     { name = "William P. Krekelberg", email = "wpk@nist.gov" },
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 classifiers = [
     "Development Status :: 2 - Pre-Alpha",
     "Intended Audience :: Science/Research",
@@ -25,7 +25,6 @@ classifiers = [
     "Operating System :: OS Independent",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -37,11 +36,14 @@ dynamic = [
     "version",
 ]
 dependencies = [
+    "attrs",
+    "simple-term-menu",
     "typer",
+    "typing-extensions; python_version<'3.11'",
 ]
 urls.Documentation = "https://pages.nist.gov/uv-workon/"
 urls.Homepage = "https://github.com/wpk-nist-gov/uv-workon"
-scripts.uv_workon = "uv_workon.cli:main"
+scripts.uvw = "uv_workon.cli:app_typer"
 
 [dependency-groups]
 dev = [
@@ -76,7 +78,7 @@ docs-live = [
     "sphinx-autobuild",
     { include-group = "docs" },
 ]
-mypy = [ "mypy>=1.14.0" ]
+mypy = [ "mypy>=1.15.0" ]
 nbval = [ "nbval" ]
 test-extras = [
     "pytest",
@@ -147,7 +149,6 @@ exclude_also = [
 ]
 omit = [
     "*/vendored/*",
-    "*/_typing.py",
 ]
 include = [
     "src/*",
@@ -181,9 +182,9 @@ exclude = [
 check_untyped_defs = true
 strict = true
 
-# [[tool.mypy.overrides]]
-# ignore_missing_imports = true
-# module = [  ]
+[[tool.mypy.overrides]]
+ignore_missing_imports = true
+module = [ "simple_term_menu.*" ]
 
 # [[tool.mypy.overrides]]
 # ignore_errors = true
@@ -270,7 +271,7 @@ reqs = [
     "grayskull>=2.5.3",
     "nbqa>=1.8.4",
     "pre-commit",
-    "pyright>=1.1.392",
+    "pyright>=1.1.396",
     "scriv",
     "twine>=5.0.0",
 ]
