@@ -61,8 +61,8 @@ extensions = [
     # - pretty things up?
     # "sphinx_design"
     # - myst stuff
-    "myst_nb",
-    # "myst_parser",
+    # "myst_nb",
+    "myst_parser",
     "sphinx_click",
 ]
 
